+++ crates/rhdl-fpga/src/dsp/mod.rs
@@ -1,4 +1,5 @@
 //! DSP Related Cores
 pub mod iq;
 pub mod lerp;
+pub mod mixer;
 pub mod nco;